% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Lyric Scraping Function}
\usage{
get_lyrics(link)
}
\arguments{
\item{link}{link to song lyrics page}
}
\value{
song lyrics in a list object
}
\description{
Function used by song_info to scrape lyrics and get into a list format
}
\examples{
get_lyrics('https://www.letras.com/cartola/')

}
