% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{scrape_songs}
\alias{scrape_songs}
\title{Scrape and save songs for an artist}
\usage{
scrape_songs(artist_name, artist_page, write_rds = FALSE, directory = NULL)
}
\arguments{
\item{artist_name}{Name of the artist to save file as}

\item{artist_page}{Website of the artist}
}
\value{
song lyric list
}
\description{
Function using other scrape_functions to scrape and save all Artist song lyrics
}
\examples{
scrape_songs('cartola','https://www.letras.com/cartola/')

}
