% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{song_info}
\alias{song_info}
\title{Song Information}
\usage{
song_info(a)
}
\arguments{
\item{a}{xml_node containing song name and link to song page}
}
\value{
a list with the name of the song, hyperlink to song page, list of song lyrics, divided by stanza
}
\description{
A function to get song information into a list object
}
\examples{
song_info(song_index('https://www.letras.com/cartola/')[[1]])

}
